Use specified image tags for demo
With the support for unified release artifacts we need a way for the
demo to support that as well.

Use the env var `VERSION_TAG` as the tag for all the stack images in the
demo docker-compose file. This is automatically set and exported by the
Makefile with logic honoring staging builds. If
`docker-compose.demo.yml` is used standalone without the Makefile, user
can `export VERSION_TAG=<stackversion>` or just leave undefined which
will revert to :latest.